third_party/WebKit/LayoutTests/http/tests/devtools/network/network-cyrillic-xhr-expected.txt

CONSOLE WARNING: line 31: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE MESSAGE: line 5: XHR loaded: 1
CONSOLE MESSAGE: line 5: XHR loaded: 2
Tests cyrillic xhr content is correctly loaded in inspector.

Bug 79026
http://127.0.0.1:8000/devtools/network/resources/cyrillic.html?sync
resource.content: The following line contains a word "success" written with cyrillic letters.
SUССЕSS

http://127.0.0.1:8000/devtools/network/resources/cyrillic.html?async
resource.content: The following line contains a word "success" written with cyrillic letters.
SUССЕSS
